Drain TridiagonalHelper_DDRB row-form ops into the shared homes (Phase 1.2b)
Moves the two shared row-form Householder ops out of TridiagonalHelper_DDRB and beside
their column-form counterparts; the tridiagonal-specific algorithm steps stay.

- computeW_Row -> Householder_DDRB (beside computeW_Column; row-form WY builder).
- innerProdRowSymm -> InnerHouseholder_DDRB (beside innerProdRow/innerProdCol).
- TridiagonalHelper_DDRB: removed both; multA_u now calls InnerHouseholder_DDRB.innerProdRowSymm;
  dropped the now-unused CommonOps_DDRM import. Retains the tridiagonal trailing-update algorithm
  (tridiagUpperRow, computeV_blockVector, applyReflectorsToRow, computeY, multA_u, computeRowOfV).
- TridiagonalDecompositionHouseholder_DDRB.getQ calls Householder_DDRB.computeW_Row.
- Tests mirror the homes: computeW_Row test moved into TestHouseholder_DDRB; removed from
  TestTridiagonalHelper_DDRB (dropped its orphaned GenericMatrixOps_F64 import).

Both ops are non-concurrent. Serial + _MT block decomposition/solver tests green (ddense + fdense).
